Cache children names in directory dentry on getdents64(2).
When the gofer mount has an authoritative cache (when the remote
filesystem is not changing under our feet), it caches all dirents
in the directory dentry on a successful getdents64(2) application
syscall. This change also additionally caches all the children
names in the dentry.

This is so that on walks to non-existent files, if the dirents
cache is available, we can return ENOENT without having to make
the round trip to the gofer.

This is especially helpful for applications that exhaustively
search all directories in a certain environment variable for
loading code. For example, $LOAD_PATH and $CLASSPATH is used
respectively by Ruby and Java.

This is more memory intensive. For sure.

Ruby example on KVM platform sees ~10% reduction in load time.
Before:
```
Finished in 1.39 seconds (files took 5.91 seconds to load)
```

After:
```
Finished in 1.23 seconds (files took 5.24 seconds to load)
```
PiperOrigin-RevId: 453362471