Deflake MsgqueueTest.MsgCtlIpcStat test.
This test flakes with <1% rate. There is a relevant comment
explaining that "timestamps only have a resolution of seconds"
and so we need to sleep for a second to actually see the timestamps
change. However, currently the tests sleeps *before* calculating
local time. Which beats the point. Move the local time declaration
before the sleep.

absl::ToTimeT() rounds down to the nearest second, discarding
nanoseconds field completely. I think the host kernel might be
rounding up the `msqid_ds` time fields.

Tested: 1000 runs of this test pass natively.
PiperOrigin-RevId: 455044862